pyproject.toml
[tool.poetry]
name = "rss-parser"
version = "0.1.0"
description = "Simple RSS parser microservice written in Python"
authors = ["Ozan Yetkin <ozan.yetkin.579@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
feedparser = "^6.0.11"
pyrebase5 = "^5.0.1"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
python-dotenv = "^1.0.1"
setuptools = "^70.3.0"
pyrebase4 = "^4.8.0"
lxml = "^5.2.2"
apscheduler = "^3.10.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
